feat: notify module

+++ src/menu-image/menu-image.service.ts
@@ -1,15 +1,16 @@
-import { BadRequestException, Injectable } from "@nestjs/common";
+import { Injectable } from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
 import { MenuImageEntity } from "./entities/menu-image.entity";
 import { Repository } from "typeorm";
 import { promises as fsPromises } from "fs";
 import { join } from "path";
-import axios from "axios";
+import { NotifyService } from "src/notify/notify.service";
 
 @Injectable()
 export class MenuImageService {
   constructor(
     @InjectRepository(MenuImageEntity) private menuImageRepository: Repository<MenuImageEntity>,
+    private notifyService: NotifyService,
   ) {}
 
   findAll() {
@@ -24,19 +25,17 @@ export class MenuImageService {
     const menuImages = files.map((file) => {
       return this.menuImageRepository.create({ url: file.path });
     });
-    const message = `Đã có món mới hôm nay, vào đặt cơm bạn nhé!`;
-    await this.sendNotifyToTelegram(message);
+    await this.notifyService.notifyNewFood();
     return this.menuImageRepository.save(menuImages);
   }
 
-  async delete({ ids, isOff }: { ids: number[]; isOff: boolean }) {
-    if (isOff) {
-      const message = `Thông báo: hôm nay quán cơm nghỉ nhé mọi người ~~`;
-      await this.sendNotifyToTelegram(message);
-    }
+  async delete(ids: number[]) {
+    if (!ids || ids.length === 0) return;
     for (const id of ids) {
-      const { url } = await this.findOne(id);
-      await this.deleteFile(url);
+      const image = await this.findOne(id);
+      if (image?.url) {
+        await this.deleteFile(image.url);
+      }
     }
     return this.menuImageRepository.delete(ids);
   }
@@ -49,20 +48,6 @@ export class MenuImageService {
       return { message: "File deleted successfully" };
     } catch (err) {
       throw new Error("Error deleting file");
-    }
-  }
-
-  async sendNotifyToTelegram(message: string) {
-    const url = `https://api.telegram.org/bot${process.env.TELEGRAM_BOT_TOKEN}/sendMessage`;
-    const payload = {
-      chat_id: process.env.TELEGRAM_CHAT_ID,
-      text: message,
-    };
-    try {
-      await axios.post(url, payload);
-      return "Gửi thông báo thành công";
-    } catch (error) {
-      throw new BadRequestException("Lỗi khi gửi thông báo");
     }
   }
 }
